Fixed  AimpVisualData

+++ AIMP.SDK/Visuals/AimpVisualData.cs
@@ -1,11 +1,13 @@
 ﻿namespace AIMP.SDK.Visuals
 {
+    using System;
+
     public class AimpVisualData
     {
         public float[] Peaks { get; set; }
 
-        public byte[][] Spectrum { get; set; }
+        public float[][] Spectrum { get; set; }
 
-        public byte[][] WaveForm { get; set; }
+        public float[][] WaveForm { get; set; }
     }
 }
